requirements.txt
# CHAP Platform - Climate and Health Analysis
chap-core>=1.0.17

# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
tensorflow>=2.13.0
torch>=2.0.0

# Geospatial Analysis
geopandas>=0.13.0
shapely>=2.0.0
fiona>=1.9.0
rasterio>=1.3.0
pyproj>=3.5.0
folium>=0.14.0
contextily>=1.3.0

# Climate Data Processing
netCDF4>=1.6.0
xarray>=2023.1.0
cftime>=1.6.0
h5py>=3.8.0

# Time Series Analysis
statsmodels>=0.14.0
pmdarima>=2.0.0
prophet>=1.1.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
streamlit>=1.28.0
dash>=2.10.0
dash-bootstrap-components>=1.4.0

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymongo>=4.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.2.0
openpyxl>=3.1.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
hypothesis>=6.75.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Logging and Monitoring
loguru>=0.7.0
python-json-logger>=2.0.0

# Notebook Support
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
